Java code that plays Tic Tac Toe
The code was designed so that you were able to play Tic Tac Toe vs. another player on the same computer, however it is very basic. The code utilizes arrays to visualize the grid of Tic Tac Toe, and uses (X, Y) coordinates to input each player's commands at a time. By inputing "0 0", you are able to select the coordinates (0,0) and place your mark on the grid. The grid will update each turn allowing the next player to see and determine their next move. If a selection is made that another player has used, then it will ask again to input another coordinate.